--- a/ast.go
+++ b/ast.go
@@ -15,6 +15,11 @@ type Expression interface {
 	ExpressionDump(w io.Writer, indent string, level int)
 }
 
+type SelectorExpression interface {
+	GetSelector() Selector
+	SetSelector(sel Selector)
+}
+
 type UnaryOperator int
 
 const (
@@ -43,6 +48,33 @@ func (op BinaryOperator) String() string {
 		return "And"
 	case BinaryOpOr:
 		return "Or"
+	default:
+		return "UNKNOWN"
+	}
+}
+
+type CollectionBindMode int
+
+const (
+	CollectionBindDefault CollectionBindMode = iota
+	CollectionBindIndex
+	CollectionBindValue
+	CollectionBindIndexAndValue
+)
+
+type CollectionOperator int
+
+const (
+	CollectionOpAny CollectionOperator = iota
+	CollectionOpAll
+)
+
+func (op CollectionOperator) String() string {
+	switch op {
+	case CollectionOpAny:
+		return "any"
+	case CollectionOpAll:
+		return "all"
 	default:
 		return "UNKNOWN"
 	}
@@ -100,6 +132,43 @@ type BinaryExpression struct {
 	Right    Expression
 }
 
+type CollectionNameBinding struct {
+	Mode    CollectionBindMode
+	Default string
+	Index   string
+	Value   string
+}
+
+func (b *CollectionNameBinding) String() string {
+	switch b.Mode {
+	case CollectionBindDefault:
+		return fmt.Sprintf("Default ( %s )", b.Default)
+	case CollectionBindIndex:
+		return fmt.Sprintf("Index ( %s )", b.Index)
+	case CollectionBindValue:
+		return fmt.Sprintf("Value ( %s )", b.Value)
+	case CollectionBindIndexAndValue:
+		return fmt.Sprintf("Index & Value ( %s, %s )", b.Index, b.Value)
+	default:
+		return "UNKNOWN"
+	}
+}
+
+type CollectionExpression struct {
+	Operator    CollectionOperator
+	Selector    Selector
+	NameBinding CollectionNameBinding
+	Expression  Expression
+}
+
+func (expr *CollectionExpression) GetSelector() Selector {
+	return expr.Selector
+}
+
+func (expr *CollectionExpression) SetSelector(sel Selector) {
+	expr.Selector = sel
+}
+
 type Selector []string
 
 func (sel Selector) String() string {
@@ -110,6 +179,14 @@ type MatchExpression struct {
 	Selector Selector
 	Operator MatchOperator
 	Value    *MatchValue
+}
+
+func (expr *MatchExpression) GetSelector() Selector {
+	return expr.Selector
+}
+
+func (expr *MatchExpression) SetSelector(sel Selector) {
+	expr.Selector = sel
 }
 
 func (expr *UnaryExpression) ExpressionDump(w io.Writer, indent string, level int) {
@@ -134,4 +211,11 @@ func (expr *MatchExpression) ExpressionDump(w io.Writer, indent string, level in
 	default:
 		fmt.Fprintf(w, "%[1]s%[3]s {\n%[2]sSelector: %[4]v\n%[1]s}\n", strings.Repeat(indent, level), strings.Repeat(indent, level+1), expr.Operator.String(), expr.Selector)
 	}
+}
+
+func (expr *CollectionExpression) ExpressionDump(w io.Writer, indent string, level int) {
+	localIndent := strings.Repeat(indent, level)
+	fmt.Fprintf(w, "%[2]s%[3]s {\n%[1]s%[2]sSelector: %[4]v\n%[1]s%[2]sNameBinding: %[5]s\n", indent, localIndent, expr.Operator.String(), expr.Selector, expr.NameBinding.String())
+	expr.Expression.ExpressionDump(w, indent, level+2)
+	fmt.Fprintf(w, "%s}\n", localIndent)
 }