--- a/bexpr.go
+++ b/bexpr.go
@@ -14,20 +14,6 @@ const (
 	defaultMaxRawValueLength = 512
 )
 
-// MatchExpressionEvaluator is the interface to implement to provide custom evaluation
-// logic for a selector. This could be used to enable synthetic fields or other
-// more complex logic that the default behavior does not support
-type MatchExpressionEvaluator interface {
-	// FieldConfigurations returns the configuration for this field and any subfields
-	// it may have. It must be valid to call this method on nil.
-	FieldConfigurations() FieldConfigurations
-
-	// EvaluateMatch returns whether there was a match or not. We are not also
-	// expecting any errors because all the validation bits are handled
-	// during parsing and cross checking against the output of FieldConfigurations.
-	EvaluateMatch(sel Selector, op MatchOperator, value interface{}) (bool, error)
-}
-
 type Evaluator struct {
 	// The syntax tree
 	ast Expression
@@ -142,12 +128,13 @@ func (eval *Evaluator) validate(config *EvaluatorConfig, dataType interface{}, u
 		maxRawValueLength = defaultMaxRawValueLength
 	}
 
-	err = validate(eval.ast, fields, config.MaxMatches, config.MaxRawValueLength)
+	ast, err := validate(eval.ast, fields, config.MaxMatches, config.MaxRawValueLength)
 	if err != nil {
 		return err
 	}
 
 	if updateEvaluator {
+		eval.ast = ast
 		eval.config = *config
 		eval.fields = fields
 		eval.boundType = rtype