This implements better support for operating on collections such as:

```
# Service.Tags is a []string
any Service.Tags as tag { tag contains "foo" }
all Service.Tags as tag { tag != "foo" }
```

This also removes the unused MatchExpressionEvaluator interface as it doesn't play nicely with the new ops. In the future we could add back an interface that allows breaking out from reflection only evaluation but for the time being this serves our purposes.