@mkeeler 2019-08-01
## Benchmark Comparison
```
ignoring BenchmarkEvaluate/Interface_Implementor/#56-8: before has 1 instances, after has 0
ignoring BenchmarkEvaluate/Interface_Implementor/#66-8: before has 1 instances, after has 0
ignoring BenchmarkGenerateFieldConfigurations/Map_of_Interface_Implementor-8: before has 1 instances, after has 0
ignoring BenchmarkEvaluate/Interface_Implementor/#10-8: before has 1 instances, after has 0
ignoring BenchmarkGenerateFieldConfigurations/Interface_Implementor-8: before has 1 instances, after has 0
ignoring BenchmarkEvaluate/Interface_Implementor/#0-8: before has 1 instances, after has 0
ignoring BenchmarkEvaluate/Interface_Implementor/#28-8: before has 1 instances, after has 0
ignoring BenchmarkEvaluate/Interface_Implementor/#38-8: before has 1 instances, after has 0
benchmark                                                          old ns/op     new ns/op     delta
BenchmarkEvaluate/Flat_Struct/#0-8                                 1077          1248          +15.88%
BenchmarkEvaluate/Flat_Struct/#1-8                                 1127          1220          +8.25%
BenchmarkEvaluate/Flat_Struct/#38-8                                1074          1295          +20.58%
BenchmarkEvaluate/Flat_Struct/#39-8                                1155          1265          +9.52%
BenchmarkEvaluate/Flat_Struct/#46-8                                1156          1342          +16.09%
BenchmarkEvaluate/Flat_Struct/#47-8                                1095          1303          +19.00%
BenchmarkEvaluate/Flat_Struct/#52-8                                1126          1369          +21.58%
BenchmarkEvaluate/Flat_Struct/#53-8                                1181          1338          +13.29%
BenchmarkEvaluate/Flat_Struct/#56-8                                1111          1325          +19.26%
BenchmarkEvaluate/Flat_Struct/#62-8                                1426          1707          +19.71%
BenchmarkEvaluate/Flat_Struct/#63-8                                1345          1693          +25.87%
BenchmarkEvaluate/Flat_Struct_Alt_Types/#0-8                       1087          1247          +14.72%
BenchmarkEvaluate/Flat_Struct_Alt_Types/#1-8                       1101          1229          +11.63%
BenchmarkEvaluate/Flat_Struct_Alt_Types/#38-8                      1141          1264          +10.78%
BenchmarkEvaluate/Flat_Struct_Alt_Types/#39-8                      1080          1258          +16.48%
BenchmarkEvaluate/Flat_Struct_Alt_Types/#46-8                      1143          1263          +10.50%
BenchmarkEvaluate/Flat_Struct_Alt_Types/#47-8                      1134          1274          +12.35%
BenchmarkEvaluate/Flat_Struct_Alt_Types/#52-8                      1133          1286          +13.50%
BenchmarkEvaluate/Flat_Struct_Alt_Types/#53-8                      1154          1274          +10.40%
BenchmarkEvaluate/map[string]map[string]bool/#8-8                  1077          1242          +15.32%
BenchmarkEvaluate/map[string]map[string]bool/#13-8                 1212          1344          +10.89%
BenchmarkEvaluate/Nested_Structs_and_Maps/#0-8                     1339          1489          +11.20%
BenchmarkEvaluate/Nested_Structs_and_Maps/#1-8                     1300          1512          +16.31%
BenchmarkEvaluate/Nested_Structs_and_Maps/#5-8                     1625          1813          +11.57%
BenchmarkEvaluate/Nested_Structs_and_Maps/#8-8                     1341          1534          +14.39%
BenchmarkEvaluate/Nested_Structs_and_Maps/#10-8                    2221          2455          +10.54%
BenchmarkEvaluate/Nested_Structs_and_Maps/#12-8                    1931          2130          +10.31%
BenchmarkGenerateFieldConfigurations/Flat_Struct-8                 5622          5553          -1.23%
BenchmarkGenerateFieldConfigurations/map[string]bool-8             1240          1181          -4.76%
BenchmarkGenerateFieldConfigurations/Nested_Structs_and_Maps-8     5221          5271          +0.96%
BenchmarkFilter/Slice-8                                            2492          3186          +27.85%
BenchmarkFilter/Array-8                                            2308          3017          +30.72%
BenchmarkFilter/Map-8                                              3200          4082          +27.56%
BenchmarkExpressionParsing/Equals-8                                38737         38589         -0.38%
BenchmarkExpressionParsing/In-8                                    82331         83337         +1.22%
BenchmarkExpressionParsing/Is_Not_Empty-8                          62152         64500         +3.78%
BenchmarkExpressionParsing/Not_Equals_And_Equals-8                 213118        207623        -2.58%
BenchmarkExpressionParsing/Lots_of_Ops-8                           200602        213457        +6.41%
BenchmarkExpressionParsing/Many_Ors-8                              178934        194592        +8.75%
BenchmarkExpressionParsing/Lots_of_Parens-8                        33550112      36149519      +7.75%
BenchmarkExpressionParsing/Contains-8                              51361         50915         -0.87%
BenchmarkExpressionParsing/In_And_Not_Equals-8                     92589         95168         +2.79%
BenchmarkExpressionParsing/Matches-8                               51418         50468         -1.85%
BenchmarkExpressionParsing/Not_Matches-8                           59352         57169         -3.68%
BenchmarkExpressionParsing/Big_Selectors-8                         118868        122931        +3.42%
BenchmarkExpressionParsing/Not_Equals-8                            41476         42021         +1.31%
BenchmarkExpressionParsing/Not_In-8                                90890         90804         -0.09%
BenchmarkExpressionParsing/Not_Contains-8                          51068         50212         -1.68%
BenchmarkExpressionParsing/Is_Empty-8                              56263         58287         +3.60%
BenchmarkExpressionParsing/Not_In_Or_Equals-8                      88449         85701         -3.11%
BenchmarkNilRegistry-8                                             4060          4627          +13.97%
BenchmarkRegistry-8                                                31.2          31.9          +2.24%
```